fix: update git include test to work with any DOTFILES_DIR path
The test was searching for '.dotfiles/git/.gitconfig' but in CI the
actual path is '/Users/runner/work/dotfiles/dotfiles/git/.gitconfig'.

Change test to search for 'git/.gitconfig' which is present in all
include directives regardless of the absolute path prefix.

This makes the test work with any DOTFILES_DIR location.